+++ pages/signup.tsx
@@ -133,7 +133,6 @@ export default function SignUp() {
       }),
     },
   ];
-  console.log(formState);
 
   return (
     <AuthenticateContainer>
